requirements.txt
python-dotenv==1.1.1
qbittorrent-api==2025.7.0
loguru==0.7.3
PyYAML==6.0.2